- Added custom room painter which is able to render a arbitrary list of walls

+++ lib/models/Wall.dart
@@ -2,8 +2,8 @@ import 'package:quiver/core.dart';
 
 class Wall {
 
-  final int xStart, yStart;
-  final int xEnd, yEnd;
+  final double xStart, yStart;
+  final double xEnd, yEnd;
 
   Wall(this.xStart, this.yStart, this.xEnd, this.yEnd);
 
@@ -13,5 +13,5 @@ class Wall {
       xEnd == o.xEnd &&
       yEnd == o.yEnd;
 
-  int get hashCode => hash4(xStart.hashCode, yStart.hashCode,xEnd.hashCode,yEnd.hashCode);
+  int get hashCode => hash4(xStart.hashCode,yStart.hashCode,xEnd.hashCode,yEnd.hashCode);
 }
